[homekit] Make Thermostat.CurrentHeatingCoolingMode optional (#17191)
Just default it to OFF if it's not linked. The experience in Home
app is still decent - it says "Cool to X", "Heat to X", or
"Set to X", instead of "Cooling to X", "Heating to X" if it was
actually running. I thought about defaulting it to "HEAT" or "COOL"
depending on the current target mode, but if it's set to "AUTO",
that's not a valid state for the current mode.

Signed-off-by: Cody Cutrer <cody@cutrer.us>